# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateDocumentRequest": {
        "description": "Запит на створення документа.",
        "properties": {
          "projectId": {
            "description": "Проєкт.",
            "format": "int32",
            "type": "integer"
          },
          "sheetDefIds": {
            "description": "Аркуші, які входять у документ.",
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": "array"
          },
          "templateVersionId": {
            "description": "Опублікована версія шаблону.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "projectId",
          "templateVersionId",
          "sheetDefIds"
        ],
        "type": "object"
      },
...
      "DocumentIdResponse": {
        "description": "Створений документ.",
        "properties": {
          "documentId": {
            "description": "Ідентифікатор.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "documentId"
        ],
        "type": "object"
      },
...
      "IFormFile": {
        "format": "binary",
        "type": "string"
      },
...
      "PagedResultOfDocumentSummary": {
        "description": "Сторінка результатів. Ендпоінтів, що повертають «усе», не існує —\r\nперевіряється архітектурним тестом.",
        "properties": {
          "items": {
            "description": "Елементи сторінки.",
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Курсор наступної сторінки; `null` — кінець.",
            "type": [
              "null",
              "string"
            ]
          },
          "totalCount": {
            "description": "Загальна кількість; `null`, якщо підрахунок дорогий.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ecr.Api | v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "description": "Зведений стан документа <b>рахується запитом</b> із `wf.ApprovalState`,\r\nа не зберігається полем (`D-93`): скалярний статус був би другим\r\nджерелом істини і рано чи пізно показав би `Approved` на документі\r\nз половиною аркушів у `Draft`.",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "periodKey",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfDocumentSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfDocumentSummary"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfDocumentSummary"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Перелік документів. Право `Document.View`.",
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIdResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Створює документ. Право `Document.Create`.",
        "tags": [
          "Documents"
        ]
      }
    },
...
    "/api/v1/documents/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "periodKey",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Документ і його аркуші. Право `Document.View`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/approve": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSheetRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSheetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSheetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Погодження або відхилення аркуша.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/export": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "summary": "Експорт у `.xlsx`. Право `Document.Export`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/export/{exportId}": {
      "get": {
        "description": "⚠ Ідентифікатор експорту приходить у повідомленні прогресу задачі:\r\nсаме тому побудова повертає `202` з `jobId`, а не файл.\r\nКнига живе годину — довше тримати немає сенсу, це знімок даних на\r\nмомент побудови.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "exportId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Віддає побудовану книгу. Право `Document.Export`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/import/apply": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplyRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatchCellsResponse"
                }
              }
            },
            "description": "OK"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Застосування раніше переглянутого імпорту. Право `Document.Import`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/import/preview": {
      "post": {
        "description": "Імпорт **завжди** через перегляд diff (ФВ-4.3): застосування — окремим викликом.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Попередній перегляд імпорту. Право `Document.Import`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/recalculate": {
      "post": {
        "description": "Довга операція — у фон із прогресом; повертає `jobId`, а не результат.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPeriodRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPeriodRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPeriodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecalculationAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecalculationAcceptedResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecalculationAcceptedResponse"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "summary": "Перерахунок документа. Право `Calculation.Recalculate`.",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{id}/reopen": {
      "post": {
        "description": "Відкриття документа при <b>закритому періоді</b> відхиляється\r\n(`ECR-PRD-4223`): спершу відкривають період, і це інше право\r\n(`D-67`).",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenDocumentRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenDocumentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReopenDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Відкриває поданий документ. Право `Document.Reopen`.",
        "tags": [
          "Documents"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}